correct a mistake on loading "midi" files

+++ LR3_WMIX(720p)/Assets/Scripts/FluidManager.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Runtime.InteropServices;
 #if UNITY_5_3_OR_NEWER
 using UnityEngine;
-// using UnityEngine.Internal;
 using AudioSample = System.Single;
 #elif GODOT
 using Godot;
@@ -18,8 +17,9 @@ public unsafe static class FluidManager{
     private static UIntPtr settings = UIntPtr.Zero;
     private static UIntPtr synth = UIntPtr.Zero;
     private static bool ready = false;
-    public static int channels = 0;
-    private static int audio_period_size = 0;
+    private static int frames = 0;
+    public const int channels = 2;
+    public static int frequency { get; private set; } = 0;
     public static bool inThread = true;
     private enum PlayerStatus : byte{
         READY,    // Player is ready
@@ -33,9 +33,10 @@ public unsafe static class FluidManager{
 #else
     private const string format = "fluid_synth_write_float";
 #endif
-#if UNITY_EDITOR_LINUX || UNITY_STANDALONE_LINUX || GODOT_X11
+#if UNITY_EDITOR_LINUX || UNITY_STANDALONE_LINUX || GODOT_X11 || GODOT_LINUXBSD
     private delegate int FluidSynthWriteFunc(UIntPtr synth, int len, void* lout, int loff, int lincr, void* rout, int roff, int rincr);
     private delegate int FluidSettingsGetintFunc(UIntPtr settings, string name, out int val);
+    private delegate int FluidSettingsGetnumFunc(UIntPtr settings, string name, out double val);
     private static class V2{
         private const string PluginName = "libfluidsynth.so.2";
         [DllImport(PluginName)] public extern static UIntPtr new_fluid_settings();
@@ -44,6 +45,10 @@ public unsafe static class FluidManager{
             UIntPtr synth, string sfpath, int reset_presets);
         [DllImport(PluginName)] public extern static int fluid_settings_getint(
             UIntPtr settings, string name, out int val);
+        [DllImport(PluginName)] public extern static int fluid_settings_getnum(
+            UIntPtr settings, string name, out double val);
+        [DllImport(PluginName)] public extern static int fluid_settings_setint(
+            UIntPtr settings, string name, int val);
         [DllImport(PluginName)] public extern static int fluid_settings_setnum(
             UIntPtr settings, string name, double val);
         [DllImport(PluginName)] public extern static void delete_fluid_synth(UIntPtr synth);
@@ -69,6 +74,10 @@ public unsafe static class FluidManager{
             UIntPtr synth, string sfpath, int reset_presets);
         [DllImport(PluginName)] public extern static int fluid_settings_getint(
             UIntPtr settings, string name, out int val);
+        [DllImport(PluginName)] public extern static int fluid_settings_getnum(
+            UIntPtr settings, string name, out double val);
+        [DllImport(PluginName)] public extern static int fluid_settings_setint(
+            UIntPtr settings, string name, int val);
         [DllImport(PluginName)] public extern static int fluid_settings_setnum(
             UIntPtr settings, string name, double val);
         [DllImport(PluginName)] public extern static void delete_fluid_synth(UIntPtr synth);
@@ -90,6 +99,8 @@ public unsafe static class FluidManager{
     private static Func<UIntPtr,UIntPtr> new_fluid_synth = null;
     private static Func<UIntPtr,string,int,int> fluid_synth_sfload = null;
     private static FluidSettingsGetintFunc fluid_settings_getint = null;
+    private static FluidSettingsGetnumFunc fluid_settings_getnum = null;
+    private static Func<UIntPtr,string,int,int> fluid_settings_setint = null;
     private static Func<UIntPtr,string,double,int> fluid_settings_setnum = null;
     private static Action<UIntPtr> delete_fluid_synth = null;
     private static Action<UIntPtr> delete_fluid_settings = null;
@@ -110,6 +121,8 @@ public unsafe static class FluidManager{
             new_fluid_synth = V2.new_fluid_synth;
             fluid_synth_sfload = V2.fluid_synth_sfload;
             fluid_settings_getint = V2.fluid_settings_getint;
+            fluid_settings_getnum = V2.fluid_settings_getnum;
+            fluid_settings_setint = V2.fluid_settings_setint;
             fluid_settings_setnum = V2.fluid_settings_setnum;
             delete_fluid_synth = V2.delete_fluid_synth;
             delete_fluid_settings = V2.delete_fluid_settings;
@@ -129,6 +142,8 @@ public unsafe static class FluidManager{
             new_fluid_synth = V3.new_fluid_synth;
             fluid_synth_sfload = V3.fluid_synth_sfload;
             fluid_settings_getint = V3.fluid_settings_getint;
+            fluid_settings_getnum = V3.fluid_settings_getnum;
+            fluid_settings_setint = V3.fluid_settings_setint;
             fluid_settings_setnum = V3.fluid_settings_setnum;
             delete_fluid_synth = V3.delete_fluid_synth;
             delete_fluid_settings = V3.delete_fluid_settings;
@@ -159,6 +174,10 @@ public unsafe static class FluidManager{
         UIntPtr synth, string sfpath, int reset_presets);
     [DllImport(PluginName)] private extern static int fluid_settings_getint(
         UIntPtr settings, string name, out int val);
+    [DllImport(PluginName)] private extern static int fluid_settings_getnum(
+        UIntPtr settings, string name, out double val);
+    [DllImport(PluginName)] private extern static int fluid_settings_setint(
+        UIntPtr settings, string name, int val);
     [DllImport(PluginName)] private extern static int fluid_settings_setnum(
         UIntPtr settings, string name, double val);
     [DllImport(PluginName)] private extern static void delete_fluid_synth(UIntPtr synth);
@@ -177,17 +196,22 @@ public unsafe static class FluidManager{
     // [DllImport(PluginName)] private extern static int fluid_player_seek(UIntPtr player, int ticks);
 #endif
     public static void Init(string sfpath, double gain = double.NaN, double overflow_vol = double.NaN){
-        if(!File.Exists(sfpath)){ ready = false; return; }
+        if((Func<UIntPtr>)new_fluid_settings == null || !File.Exists(sfpath)){ ready = false; return; }
         CleanUp();
         settings = new_fluid_settings();
         if(settings == UIntPtr.Zero){ ready = false; return; }
         synth = new_fluid_synth(settings);
         if(synth == UIntPtr.Zero) goto cleanup;
         if(fluid_synth_sfload(synth, sfpath, 0) == -1) goto cleanup;
-        fluid_settings_getint(settings, "synth.audio-channels", out channels);
-        if(channels < 1) goto cleanup;
-        fluid_settings_getint(settings, "audio.period-size", out audio_period_size);
-        if(audio_period_size < channels) goto cleanup;
+        fluid_settings_getnum(settings, "synth.sample-rate", out double sampleRate);
+        frequency = (int)sampleRate;
+        if(frequency < 8000) goto cleanup;
+        // fluid_settings_getint(settings, "audio.periods", out int audio_periods);
+        // if(audio_periods < 2) goto cleanup;
+        fluid_settings_getint(settings, "audio.period-size", out int audio_period_size);
+        if(audio_period_size < 64) goto cleanup;
+        frames = audio_period_size / channels;
+        fluid_settings_setint(settings, "synth.audio-channels", channels / 2);
         if(!double.IsNaN(overflow_vol) && !double.IsInfinity(overflow_vol) && overflow_vol >= double.Epsilon)
             fluid_settings_setnum(settings, "synth.overflow.volume", overflow_vol);// default:500
         if(!double.IsNaN(gain) && !double.IsInfinity(gain) && gain >= double.Epsilon)
@@ -195,21 +219,20 @@ public unsafe static class FluidManager{
         ready = true; return;
         cleanup: CleanUp(); return;
     }
-    public static AudioSample[] MidiToSamples(string midipath, out int lengthSamples, out int frequency){
-        lengthSamples = frequency = 0;
+    public static AudioSample[] MidiToSamples(string midipath){
         if(!ready || !File.Exists(midipath)) return null;
         UIntPtr player = new_fluid_player(synth);
         if(player == UIntPtr.Zero) return null;
         List<AudioSample> total_samples = new List<AudioSample>();
 #if GODOT
-        byte[] samples = new byte[audio_period_size * sizeof(short)];
+        byte[] samples = new byte[frames * channels * sizeof(short)];
 #else
-        float[] samples = new float[audio_period_size];
+        float[] samples = new float[frames * channels];
 #endif
         fluid_player_add(player, midipath);
         fixed(void* temp = samples){
             /*fluid_player_set_playback_callback(player, (data, e) => {
-                if(fluid_synth_write(synth, audio_period_size, temp, 0, channels, temp, 1, channels) == 0){
+                if(fluid_synth_write(synth, frames, temp, 0, channels, temp, 1, channels) == 0){
                     total_samples.AddRange(samples);
                 }
                 return 0;
@@ -217,39 +240,25 @@ public unsafe static class FluidManager{
             fluid_player_play(player);
             while(inThread && 
                 fluid_player_get_status(player) == PlayerStatus.PLAYING &&
-                fluid_synth_write(synth, audio_period_size, temp, 0, channels, temp, 1, channels) == 0
-            ){
-#if GODOT
-                #warning why "System.RankException" in GODOT?
-#endif
-                total_samples.AddRange(samples);
-
-            }
-            fluid_player_stop(player);
+                fluid_synth_write(synth, frames, temp, 0, channels, temp, 1, channels) == 0
+            ){ total_samples.AddRange(samples); }
         }
+        fluid_player_stop(player);
         fluid_player_join(player);
         delete_fluid_player(player);
-        lengthSamples = total_samples.Count;
-#if UNITY_EDITOR_WIN || UNITY_STANDAONE_WIN
-        frequency = lengthSamples / audio_period_size * 2;
-#elif GODOT //_WINDOWS
-        frequency = lengthSamples / audio_period_size;
-// #elif GODOT_X11
-//         frequency = lengthSamples / audio_period_size / sizeof(short);
-#else
-        frequency = lengthSamples / audio_period_size / sizeof(float);
-#endif
         return total_samples.ToArray();
     }
     public static void CleanUp(){
+        frequency = frames = 0;
+        ready = false;
+        // inThread = false;
+        if((Action<UIntPtr>)delete_fluid_synth == null) return;
         if(synth != UIntPtr.Zero) delete_fluid_synth(synth);
         if(settings != UIntPtr.Zero) delete_fluid_settings(settings);
         settings = synth = UIntPtr.Zero;
-        ready = false;
-        // inThread = false;
     }
     static FluidManager(){
-#if UNITY_EDITOR_LINUX || UNITY_STANDALONE_LINUX || GODOT_X11
+#if UNITY_EDITOR_LINUX || UNITY_STANDALONE_LINUX || GODOT_X11 || GODOT_LINUXBSD
         MatchVersion();
 #endif
 // #if GODOT
